terra_graphics_engine::engine::gl2d
Module shapes

Shapes Module

This submodule of the supermodule ‘gl2d’ handles the drawing of simple 2d shapes(like rectangle, circle, etc.) and lines using openGL Immediate mode rendering.

These functions are glow implementations of the functions in engine.py.

NOTE: Some of the PyOpengl functions which are depreciated are not available in the glow crate like matrix Mode setting, or it might be available with different name, so you have to check that before the conversion. But other than most of the functions in PyOpenGl also have a similar implementation in glow like the glEnable(GL_BLEND) in PyopenGl is gl.enable(BLEND).

Functions

__pyfunction_tengine_draw_circle 🔒 ⚠

__pyfunction_tengine_draw_quarter_circle 🔒 ⚠

__pyfunction_tengine_draw_rect 🔒 ⚠

__pyfunction_tengine_draw_rounded_rect 🔒 ⚠

register_shapes_module

Declaration of the shapes submodule for the parent python module. Reference: https://pyo3.rs/v0.23.4/module.html

tengine_draw_circle

Draw a circle with the given radius at given screen coordinates(x and y).

tengine_draw_quarter_circle

Draw a quarter circle on the screen with rounding based on the provided factor.The roundedness does not affect the radius.:param x: The x-coordinate of the corner of the quarter circle.:param y: The y-coordinate of the corner of the quarter circle.:param radius: The radius of the quarter circle.:param color: The color (RGBA) of the quarter circle.:param segments: The number of segments for drawing the curve.:param corner: The corner for the quarter circle (‘top-left’, ‘top-right’, ‘bottom-left’, ‘bottom-right’).:param roundedness: The factor of rounding (0=sharp, 1=circle).

tengine_draw_rect

Draw a rectangle of the given width and height at given screen coordinates(x and y), of color [r: f32, g: f32, b: f32, a: f32].

tengine_draw_rounded_rect

Draw a rectangle with rounded corners using four quarter circles for each corner:param x: The x-coordinate of the rectangle.:param y: The y-coordinate of the rectangle.:param color: The color (RGBA) of the rectangle.:param radius: The radius of the quarter circle corners of the rectangle.