terra_graphics_engine::engine::gl2d::shapes
Function tengine_draw_circle

pub fn tengine_draw_circle(
    x: f32,
    y: f32,
    radius: f32,
    segments: u32,
    color: [f32; 4],
)

Draw a circle with the given radius at given screen coordinates(x and y).

The circle is approximated using a polygon with many segments. And since it is being approximated( that is made up of n number of polygons) then setting the value of this n(segments) will allow you to create any polygon, like hexagon( n = 6 ) or octagon( n = 8 )!