terra_graphics_engine::engine
Module test

Test Module

Used to test functionality of some engine functions/methods be calling a python function.

Functions

__pyfunction_draw_color_changing_shapes 🔒 ⚠

__pyfunction_run_color_gradient 🔒 ⚠

draw_color_changing_shapes

Draws different types of shapes like rectangle, hexagon, and a circle that change there color every frame.

hsv_to_rgb 🔒

Converts an HSV color to its RGB equivalent.

register_test_module

Declaration of the test submodule for the parent python module. Reference: https://pyo3.rs/v0.23.4/module.html

run_color_gradient

A simple color gradient sdl3 example but using OpenGL(Glow crate).