terra_graphics_engine::engine
Module window

Window Module for Terra Graphics Engine.

This module is used to initialise program windows(both sdl based and gl based) using the sdl3 context and video_subsystem.

Functions

__pyfunction_init_gl_window 🔒 ⚠

__pyfunction_set_window_icon 🔒 ⚠

init_gl_window

Create SDL3 Window with OpenGL context.

init_opengl_drawing

Initialise the Global vao, vbo and shader program and also sets gl_viewport to wndow size.

register_window_module

Declaration of the window submodule for the parent python module. Reference: https://pyo3.rs/v0.23.4/module.html

set_window_icon

Loads the icon from the relative icon path passed, and sets it as the current window icon.