terra_graphics_engine
Struct TContext

pub struct TContext {
    pub(crate) gl_context: Rc<GLContext>,
    pub(crate) sdl_context: Sdl,
    pub(crate) sdl_video_subsystem: VideoSubsystem,
    pub(crate) sdl_window: Window,
    pub(crate) global_vao: RwLock<Option<NativeVertexArray>>,
    pub(crate) global_vbo: RwLock<Option<NativeBuffer>>,
    pub(crate) global_program: RwLock<Option<NativeProgram>>,
    pub(crate) shader_programs: RwLock<HashMap<String, NativeProgram>>,
}

NOTE: I am for now wrapping shader_program, vao, and vbo in rwlock so that we can mutate it later, after the initialisation of the global context.

Remember that Locking to mutate variable in RWLOCK is very expensive and is slow compared to directly reading the value(as it uses Atomic operations) so it will be better if we initialise the global contexts only once!

Fields
gl_context: Rc<GLContext>sdl_context: Sdlsdl_video_subsystem: VideoSubsystemsdl_window: Windowglobal_vao: RwLock<Option<NativeVertexArray>>global_vbo: RwLock<Option<NativeBuffer>>global_program: RwLock<Option<NativeProgram>>shader_programs: RwLock<HashMap<String, NativeProgram>>
Implementations

impl TContext

pub fn init( sdl_ctx: Sdl, gl_ctx: GLContext, sdl_vs_ctx: VideoSubsystem, window_ctx: Window, ) -> TContext

pub fn get_all(&self) -> (EventPump, &Window, Context)

pub fn get_window(&self) -> &Window

pub fn get_event(&self) -> EventPump

pub fn get_gl(&self) -> Context

pub fn make_gl_current(&self)

pub fn get_gl_parameters( &self, ) -> (NativeVertexArray, NativeBuffer, NativeProgram)

Returns the current Global gl_Parameters(vao, vbo and program) from the T_Context Object(Immutable).

pub fn get_glob_vao(&self) -> NativeVertexArray

Returns the current Global Vertex_Array from the T_Context Object(Immutable).

pub fn get_glob_vbo(&self) -> NativeBuffer

Returns the current Global Vertex_Buffer from the T_Context Object(Immutable).

pub fn get_glob_glprogam(&self) -> NativeProgram

Returns the current Global Shader Program from the T_Context Object(Immutable).

pub fn create_shader_program( &self, gl: &Context, program_name: String, vertex_source: &str, fragment_source: &str, )

pub fn set_shader_program(&self, gl: &Context, program_name: String)

pub fn get_shader_program(&self, program_name: String) -> NativeProgram

pub fn add_shader_program( &self, program_name: String, shader_program: NativeProgram, )

Adds a new shader program to the global shader programs if it does not exists, if a shader program at the the given porgram_name does exist, then it just updates the value of it!

Example:

let vertex_shader = compile_shader(&gl, VERTEX_SHADER, vertex_shader_src);

let fragment_shader = compile_shader(&gl, FRAGMENT_SHADER, fragment_shader_src);

let program = link_program(&gl, vertex_shader, fragment_shader);

add_shader_program(String::from("shader_program1"), program);

Note!

The given functions gets a write(mutable) lock to the shader_programs, make sure there is no existing write lock present before calling this function or the engine will panic!